requirement.txt
channels==4.0.0
daphne==4.0.0
Django==4.2
django-environ==0.10.0
channels_redis==4.1.0
djangorestframework=3.14.0